# Optimize InlineFormatter