--- a/src/MiniProfiler.Shared/SqlFormatters/InlineFormatter.cs
+++ b/src/MiniProfiler.Shared/SqlFormatters/InlineFormatter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Text.RegularExpressions;
-using StackExchange.Profiling.Internal;
 
 namespace StackExchange.Profiling.SqlFormatters
 {
@@ -9,7 +9,6 @@ namespace StackExchange.Profiling.SqlFormatters
     /// </summary>
     public class InlineFormatter : ISqlFormatter
     {
-        private static readonly Regex ParamPrefixes = new Regex("[@:?].+", RegexOptions.IgnoreCase | RegexOptions.Compiled);
         private static bool includeTypeInfo;
 
         /// <summary>
@@ -35,22 +34,21 @@ namespace StackExchange.Profiling.SqlFormatters
                 return commandText;
             }
 
-            var originalCommandText = commandText;
-
+            var paramValuesByName = new Dictionary<string, string>(parameters.Count);
             foreach (var p in parameters)
             {
-                // If the parameter doesn't have a prefix (@,:,etc), append one
-                var name = ParamPrefixes.IsMatch(p.Name)
-                    ? p.Name
-                    : Regex.Match(originalCommandText, "([@:?])" + Regex.Escape(p.Name), RegexOptions.IgnoreCase).Value;
-                if (name.HasValue())
-                {
-                    var value = GetParameterValue(p);
-                    commandText = Regex.Replace(commandText, "(" + Regex.Escape(name) + ")([^0-9A-z]|$)", m => value + m.Groups[2], RegexOptions.IgnoreCase);
-                }
+                var trimmedName = p.Name.TrimStart('@', ':', '?').ToLower();
+                paramValuesByName[trimmedName] = GetParameterValue(p);
             }
 
-            return commandText;
+            var regexPattern = "[@:?](?:" + string.Join("|", paramValuesByName.Keys.Select(Regex.Escape)) + ")(?![0-9a-z])";
+
+            return Regex.Replace(
+                commandText,
+                regexPattern,
+                m => paramValuesByName[m.Value.Substring(1).ToLower()],
+                RegexOptions.IgnoreCase
+            );
         }
 
         /// <summary>